convert ints back to 64 bit

+++ mcdc/kernel.py
@@ -990,7 +990,7 @@ def bank_scanning(bank, mcdc):
     N_local = bank["size"]
 
     # Starting index
-    buff = np.zeros(1, dtype=np.int32)
+    buff = np.zeros(1, dtype=np.int64)
     with objmode():
         MPI.COMM_WORLD.Exscan(np.array([N_local]), buff, MPI.SUM)
     idx_start = buff[0]
@@ -1040,7 +1040,7 @@ def bank_scanning_DNP(bank, mcdc):
         N_local += math.ceil(DNP["w"])
 
     # Starting index
-    buff = np.zeros(1, dtype=np.int32)
+    buff = np.zeros(1, dtype=np.int64)
     with objmode():
         MPI.COMM_WORLD.Exscan(np.array([N_local]), buff, MPI.SUM)
     idx_start = buff[0]
@@ -3031,13 +3031,13 @@ def scramble_LDS(mcdc):
 @njit
 def rhalton(N, dim, seed=12345, skip=0):
     np.random.seed(seed)
-    primes = np.array((2, 3, 5, 7, 11, 13, 17, 19, 23, 29), dtype=np.int32)
+    primes = np.array((2, 3, 5, 7, 11, 13, 17, 19, 23, 29), dtype=np.int64)
     halton = np.zeros((N, dim), dtype=np.float64)
 
     for D in range(dim):
         b = primes[D]
-        # b = np.int32(2)
-        ind = np.arange(skip, skip + N, dtype=np.int32)
+        # b = np.int64(2)
+        ind = np.arange(skip, skip + N, dtype=np.int64)
         b2r = 1 / b
         ans = np.zeros(ind.shape, dtype=np.float64)
         res = ind.copy()
@@ -3047,7 +3047,7 @@ def rhalton(N, dim, seed=12345, skip=0):
             pdig = perm[dig]
             ans = ans + pdig.astype(np.float64) * b2r
             b2r = b2r / np.float64(b)
-            res = ((res - dig) / b).astype(np.int32)
+            res = ((res - dig) / b).astype(np.int64)
         halton[:, D] = ans
 
     return halton
@@ -3055,7 +3055,7 @@ def rhalton(N, dim, seed=12345, skip=0):
 
 @njit
 def halton(N, dim, skip=0):
-    primes = np.array((2, 3, 5, 7, 11, 13, 17, 19, 23, 29), dtype=np.int32)
+    primes = np.array((2, 3, 5, 7, 11, 13, 17, 19, 23, 29), dtype=np.int64)
     halton = np.zeros((N, dim), dtype=np.float64)
 
     for D in range(dim):
